+++ changelog.md
@@ -8,8 +8,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Todo
+
+2025/08/02 08:29:10 Error starting transcription job: operation error Transcribe: StartTranscriptionJob, https response error StatusCode: 400, RequestID: 0fb0a4f6-5163-4cdb-8c5d-44491af38dbd, BadRequestException: The specified S3 bucket isn't in the same region. Make sure the bucket is in the eu-central-1 region and try your request again.
+
+
+## [v0.2.0]
 
 ### Added
+- Copy Result Button for resulttext
+- when called from CLI, first AWS environment variables are user for credentials
+- Bedrock model is configurable from dialog
 
 ### Changed
 - refactor: panels in own package
@@ -21,7 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- WIP: open load Dialog with saves directory
+- open load Dialog now uses saved directory
 
 ## [v0.1.0]